+++ src/Interfaces.ts
@@ -1,5 +1,10 @@
 import { Resource } from "./Resource";
-import { ConditionTypes, QueryFunctionType } from "./Types";
+import {
+  ConditionTypes,
+  QueryFunctionType,
+  RequestInterceptorType,
+  ResponseInterceptorType,
+} from "./Types";
 
 export interface IConfig {
   baseURL?: string;
@@ -13,8 +18,8 @@ export interface IInternalConfig extends IConfig {
 }
 
 export interface IInterceptors {
-  requests: any[];
-  responses: any[];
+  requests: RequestInterceptorType[];
+  responses: ResponseInterceptorType[];
 }
 
 export interface IPaginate {