+++ tests/index.spec.js
@@ -179,7 +179,9 @@ describe("axe-api-client", () => {
     expect(request.method).toBe("POST");
     expect(request.headers["Content-Type"]).toBe("application/json");
     expect(request.body).toBe(`{"name":"Karl"}`);
-    expect(response.id).toBe(100);
+    const data = await response.json();
+    expect(response.status).toBe(200);
+    expect(data.id).toBe(100);
   });
 
   test(`update()`, async () => {
@@ -197,7 +199,8 @@ describe("axe-api-client", () => {
     expect(url).toBe("https://axe-api.com/api/v1/users/1");
     expect(request.method).toBe("PUT");
     expect(request.body).toBe(`{"name":"Karl"}`);
-    expect(response.id).toBe(100);
+    const data = await response.json();
+    expect(data.id).toBe(100);
   });
 
   test(`patch()`, async () => {
@@ -215,7 +218,8 @@ describe("axe-api-client", () => {
     expect(url).toBe("https://axe-api.com/api/v1/users/1");
     expect(request.method).toBe("PATCH");
     expect(request.body).toBe(`{"name":"Karl"}`);
-    expect(response.id).toBe(100);
+    const data = await response.json();
+    expect(data.id).toBe(100);
   });
 
   test(`delete()`, async () => {
@@ -229,7 +233,8 @@ describe("axe-api-client", () => {
     expect(url).toBe("https://axe-api.com/api/v1/users/1");
     expect(request.method).toBe("DELETE");
     expect(request.body).toBe(undefined);
-    expect(response).toBe(undefined);
+    const json = await response.json();
+    expect(json).toBe(undefined);
   });
 
   test(`paginate()`, async () => {
@@ -245,7 +250,8 @@ describe("axe-api-client", () => {
     const request = global.fetch.mock.calls[0][1];
     expect(url).toBe("https://axe-api.com/api/v1/users?page=10&per_page=5");
     expect(request.method).toBe("GET");
-    expect(response).toBe(mockResponse);
+    const json = await response.json();
+    expect(json).toBe(mockResponse);
   });
 
   test(`URL Tests`, async () => {
@@ -266,7 +272,8 @@ describe("axe-api-client", () => {
       `https://axe-api.com/api/v1/users?page=1&per_page=10&fields=id&sort=id&q=%5B%7B%22id%22%3A1%7D%5D`,
     );
     expect(request.method).toBe("GET");
-    expect(response).toBe(mockResponse);
+    const json = await response.json();
+    expect(json).toBe(mockResponse);
   });
 
   test(`post()`, async () => {
@@ -282,7 +289,8 @@ describe("axe-api-client", () => {
     expect(url).toBe("https://axe-api.com/api/v1/users");
     expect(request.method).toBe("POST");
     expect(JSON.parse(request.body).name).toBe(data.name);
-    expect(response).toBe("RESULT");
+    const json = await response.json();
+    expect(json).toBe("RESULT");
   });
 
   test(`put()`, async () => {
@@ -298,7 +306,8 @@ describe("axe-api-client", () => {
     expect(url).toBe("https://axe-api.com/api/v1/users");
     expect(request.method).toBe("PUT");
     expect(JSON.parse(request.body).name).toBe(data.name);
-    expect(response).toBe("RESULT");
+    const json = await response.json();
+    expect(json).toBe("RESULT");
   });
 
   test(`patch()`, async () => {
@@ -314,6 +323,7 @@ describe("axe-api-client", () => {
     expect(url).toBe("https://axe-api.com/api/v1/users");
     expect(request.method).toBe("PATCH");
     expect(JSON.parse(request.body).name).toBe(data.name);
-    expect(response).toBe("RESULT");
+    const json = await response.json();
+    expect(json).toBe("RESULT");
   });
 });